#40 feat: 메인 페이지 범위에서 빌드 오류를 일으키는 타입 오류 수정
