+++ app/main/LoginInfoArea/index.tsx
@@ -9,9 +9,8 @@ interface Props {
 }
 
 export default function LoginInfoArea({ handleAuthenticationModal }: Props) {
-    // const { data: session } = useSession();
-    // const session = { user: { email: "" } };
-    const session = null;
+    // FIXME: 타입 오류 수정 필요
+    const { data: session } = useSession() ?? {};
 
     return (
         <section className="flex flex-col items-end gap-4 w-full md:flex-row md:gap-8 md:items-center md:justify-end">
